backend/requirements.txt
click==8.0.2
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
paho-mqtt==1.5.1
psycopg2-binary==2.9.1
python-etcd==0.4.5
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
